feat(otlp-receiver): Improve command documentation
